DBI methods

Implementations of pure virtual functions defined in the DBI package.

Usage

# S4 method for class 'polarssql_result'
dbClearResult(res, ...)

# S4 method for class 'polarssql_driver'
dbConnect(drv, ...)

# S4 method for class 'polarssql_connection'
dbDisconnect(conn, ...)

# S4 method for class 'polarssql_connection,character'
dbExistsTable(conn, name, ...)

# S4 method for class 'polarssql_result'
dbFetch(res, n = -1, ...)

# S4 method for class 'polarssql_result'
dbGetRowsAffected(res, ...)

# S4 method for class 'polarssql_result'
dbHasCompleted(res, ...)

# S4 method for class 'polarssql_connection'
dbIsValid(dbObj, ...)

# S4 method for class 'polarssql_driver'
dbIsValid(dbObj, ...)

# S4 method for class 'polarssql_result'
dbIsValid(dbObj, ...)

# S4 method for class 'polarssql_connection,character'
dbListFields(conn, name, ...)

# S4 method for class 'polarssql_connection'
dbListTables(conn, ...)

# S4 method for class 'polarssql_connection,character'
dbQuoteIdentifier(conn, x, ...)

# S4 method for class 'polarssql_connection,character'
dbQuoteString(conn, x, ...)

# S4 method for class 'polarssql_connection,character'
dbRemoveTable(conn, name, ..., fail_if_missing = TRUE)

# S4 method for class 'polarssql_connection,character'
dbSendQuery(conn, statement, ...)

# S4 method for class 'polarssql_connection,character,data.frame'
dbWriteTable(conn, name, value, ..., overwrite = FALSE)

# S4 method for class 'polarssql_connection'
show(object)

# S4 method for class 'polarssql_driver'
show(object)

# S4 method for class 'polarssql_result'
show(object)

Arguments

res

An object inheriting from DBIResult.

...

Other arguments passed on to methods.

drv

an object that inherits from DBIDriver, or an existing DBIConnection object (in order to clone an existing connection).

conn

A DBIConnection object, as returned by dbConnect().

name

The table name, passed on to dbQuoteIdentifier(). Options are:

a character string with the unquoted DBMS table name, e.g. "table_name",

a call to Id() with components to the fully qualified table name, e.g. Id(schema = "my_schema", table = "table_name")

a call to SQL() with the quoted and fully qualified table name given verbatim, e.g. SQL('"my_schema"."table_name"')

n

maximum number of records to retrieve per fetch. Use n = -1 or n = Inf to retrieve all pending records. Some implementations may recognize other special values.

dbObj

An object inheriting from DBIObject, i.e. DBIDriver, DBIConnection, or a DBIResult

x

A character vector, SQL or Id object to quote as identifier.

fail_if_missing

If FALSE, dbRemoveTable() succeeds if the table doesn't exist.

statement

a character string containing SQL.

value

A data.frame (or coercible to data.frame).

overwrite

Allow overwriting the destination table. Cannot be TRUE if append is also TRUE.

object

Any R object